rom/exec: Some more -Wall cleanups

git-svn-id: https://svn.aros.org/svn/aros/trunk@35097 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/exec/debug.c
@@ -141,7 +141,7 @@ static char *NextWord(char *s)
 	    RawPutChar(key);
 	    comm[i++] = key;
 	}
-	while (i < sizeof(comm)-1);
+	while (i < (int)sizeof(comm)-1);
 	comm[i] = 0;
 	RawPutChar('\n');
